Rename Home -> Exams, add styling, update routing with react router
