+++ client/src/App.css
@@ -5,4 +5,9 @@
   left: 0%;
 }
 
+.NotFound {
+  margin: 25% 0;
+  display: flex;
+  justify-content: center;
+}
