+++ client/src/components/Admin/Admin.js
@@ -1,9 +1,9 @@
-import { Navbar } from "./Navbar/Navbar";
+import './Admin.css';
+import { Navbar } from "../Navbar/Navbar";
 
 export const Admin = () => {
     return (
         <>
-            <Navbar />
             <p style={{marginTop: "500px"}}> TESTING ADMIN PAGE </p>
         </>
     )
