+++ client/src/components/Navbar/Navbar.css
@@ -2,6 +2,7 @@
     background-color: rgb(0, 0, 8);
     width: 100vw;
     margin: 0;
+    height: 8%;
 }
 
 .nav {
@@ -11,7 +12,7 @@
 
 .nav img {
     height: 2rem;
-    padding: 2%;
+    padding: 1.5%;
     margin-left: 5%;
 }
 
@@ -48,4 +49,8 @@
     text-decoration: none;
     color: rgb(151, 148, 148);
     font-weight:900;
+}
+
+.active {
+    text-decoration: overline !important;
 }
