+++ client/src/components/NotFound/NotFound.css
@@ -1,0 +1,5 @@
+.NotFound p {
+    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    font-size: 2rem;
+    color: rgb(0, 0, 0);
+}
