+++ client/src/components/NotFound/NotFound.js
@@ -1,0 +1,10 @@
+import './NotFound.css';
+import { Navbar } from "../Navbar/Navbar";
+
+export const NotFound = () => {
+    return (
+        <div className="NotFound">
+            <p>Sorry. Something went wrong..</p>
+        </div>
+    )
+};
